% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bufferct.R
\name{bufferct}
\alias{bufferct}
\title{Create dissolved buffers around points with point count attribute}
\usage{
bufferct(sf_pts, bufsize)
}
\arguments{
\item{sf_pts}{An \code{sf} point object to buffer.}

\item{bufsize}{Size of buffer.}
}
\value{
An \code{sf} polygon object.
}
\description{
Create dissolved buffers around points with point count attribute
}
