% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disperse_pts.R
\name{disperse_pts}
\alias{disperse_pts}
\title{Disperse points}
\usage{
disperse_pts(sf_pts, space_dist = 50000)
}
\arguments{
\item{sf_pts}{An \code{sf} point object to disperse.}

\item{space_dist}{Maximum spacing threshold of points within \code{sf_pts} to disperse. Under this threshold, neighboring points will be considered a cluster to be dispersed. Units are projecction units of \code{sf_pts}.}
}
\value{
An \code{sf} point object.
}
\description{
Disperse points
}
