% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mincircle.R
\name{mincircle}
\alias{mincircle}
\title{Return radius of minimum encompassing circle}
\usage{
mincircle(n)
}
\arguments{
\item{n}{Number of points (unit circles) to encompass.}
}
\value{
Radius of minimum encompassing circle to cover \code{n} unit circles. Below \code{n}<=20, returns exact value, based on \href{https://erich-friedman.github.io/packing/cirincir/}{theory}. Above 20, returns sqrt(\code{n}*1.35) as an approximation.
}
\description{
Return radius of minimum encompassing circle
}
